py-srv/api/bin/requirements.txt
python-multipart 
fastapi 
uvicorn 
pydantic
tortoise-orm
asyncpg
psycopg2
